refactor: Move get_current_user to dependencies
