+++ lilycloudproto/apis/admin/task.py
@@ -3,8 +3,8 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from lilycloudproto.apis.auth import get_current_user
 from lilycloudproto.database import get_db
+from lilycloudproto.dependencies import get_current_user
 from lilycloudproto.domain.entities.task import Task
 from lilycloudproto.domain.entities.user import User
 from lilycloudproto.domain.values.task import ListArgs
